✨ [FE] feat: edit 페이지 img 위로 올리기
